+++ packages/theme/src/cli/polyfills/promiseWithResolvers.ts
@@ -1,0 +1,22 @@
+interface PromiseWithResolvers<T> {
+  promise: Promise<T>
+  resolve: (value: T | PromiseLike<T>) => void
+  reject: (reason?: unknown) => void
+}
+
+// Polyfill for Promise.withResolvers
+// Can remove once our minimum supported Node version is 22
+export function promiseWithResolvers<T>(): PromiseWithResolvers<T> {
+  if (typeof Promise.withResolvers === 'function') {
+    return Promise.withResolvers<T>()
+  }
+
+  let resolve!: (value: T | PromiseLike<T>) => void
+  let reject!: (reason?: unknown) => void
+  const promise = new Promise<T>((_resolve, _reject) => {
+    resolve = _resolve
+    reject = _reject
+  })
+
+  return {promise, resolve, reject}
+}
